correction bug suppresion mail inbox
Modification de la class et du fichier inbox pour effacement correct des mails
après relève

+++ cron/inbox.php
@@ -58,36 +58,42 @@ if ($connection = $pop->pop3_login($p['config']['apicryptPopHost'], $p['config']
         foreach ($liste as $msgID=>$msgV) {
             $message=$pop->mail_mime_to_array($connection, $msgID, false);
 
-            foreach ($message as $kpart=>$part) {
-                $filename=date("Ymd-His", $msgV['udate']).'-'.$msgV['uid'];
-                $dirname=date("Ymd-His", $msgV['udate']).'-'.$msgV['uid'].'.f';
+            if (is_array($message)) {
+                foreach ($message as $kpart=>$part) {
+                    $filename=date("Ymd-His", $msgV['udate']).'-'.$msgV['uid'];
+                    $dirname=date("Ymd-His", $msgV['udate']).'-'.$msgV['uid'].'.f';
 
-                if ($kpart > 0 and isset($part['is_attachment'])) {
-                    msTools::checkAndBuildTargetDir($p['config']['apicryptCheminInbox'].$dirname);
+                    if ($kpart > 0 and isset($part['is_attachment'])) {
+                        if (strlen($part['data'])>10) {
+                            msTools::checkAndBuildTargetDir($p['config']['apicryptCheminInbox'].$dirname);
+                            $filec=$p['config']['apicryptCheminInbox'].$dirname.'/'.$part['filename'];
+                            $filenc=$p['config']['apicryptCheminInbox'].$dirname.'/';
+                            file_put_contents($filec, $part['data']);
+                            msApicrypt::decrypterPJ($filec, $filenc);
+                            unlink($filec);
+                        }
+                    } elseif ($kpart > 0) {
+                        if (strlen($part['data'])>10) {
+                            $filec=$p['config']['apicryptCheminInbox'].$filename;
+                            $filenc=$p['config']['apicryptCheminInbox'].$filename.'.txt';
 
-                    $filec=$p['config']['apicryptCheminInbox'].$dirname.'/'.$part['filename'];
-                    $filenc=$p['config']['apicryptCheminInbox'].$dirname.'/';
-                    file_put_contents($filec, $part['data']);
-                    msApicrypt::decrypterPJ($filec, $filenc);
-                    unlink($filec);
-                } elseif ($kpart > 0) {
-                    $filec=$p['config']['apicryptCheminInbox'].$filename;
-                    $filenc=$p['config']['apicryptCheminInbox'].$filename.'.txt';
+                            file_put_contents($filec, $part['data']);
+                            msApicrypt::decrypterCorps($filec, $filenc);
 
-                    file_put_contents($filec, $part['data']);
-                    msApicrypt::decrypterCorps($filec, $filenc);
-
-                    if (is_file($filenc)) {
-                      unlink($filec);
-                    } else {
-                      rename($filec, $filenc);
+                            if (is_file($filenc)) {
+                                unlink($filec);
+                            } else {
+                                rename($filec, $filenc);
+                            }
+                        }
                     }
                 }
+                // supprimer le message
+                $pop->pop3_dele($connection, $msgID);
             }
-            // supprimer les messages
-            //print_r($pop->pop3_dele($connection, $msgID));
-
         }
+        // supprimer les messages
+        $pop->pop3_expunge($connection);
     }
 }
 
@@ -103,35 +109,35 @@ foreach ($scanned_directory as $file) {
 
   //si c'est un txt
   if (substr($file, -4) == '.txt') {
-      $hprim = msHprim::getHprimHeaderData($p['config']['apicryptCheminInbox'].'/'.$file);
+    $hprim = msHprim::getHprimHeaderData($p['config']['apicryptCheminInbox'].'/'.$file);
 
-      $hprim=msTools::utf8_converter($hprim);
+    $hprim=msTools::utf8_converter($hprim);
 
-      $filedata = msInbox::getFileDataFromName($file);
+    $filedata = msInbox::getFileDataFromName($file);
 
     //pj
     $dir=$p['config']['apicryptCheminInbox'].'/'.str_ireplace('.txt', '.f', $file);
-      if (is_dir($dir)) {
-          $pj = array_diff(scandir($dir), array('..', '.'));
-      } else {
-          $pj=null;
-      }
+    if (is_dir($dir)) {
+      $pj = array_diff(scandir($dir), array('..', '.'));
+    } else {
+      $pj=null;
+    }
 
 
-      $data=array(
-      'txtFileName'=>$file,
-      'txtDatetime'=>$filedata['datetime'],
-      'txtNumOrdre'=>$filedata['numOrdre'],
-      'hprimIdentite'=>$hprim['prenom'].' '.$hprim['nom'],
-      'hprimExpediteur'=>$hprim['expediteur'],
-      'hprimCodePatient'=>$hprim['codePatient'],
-      'hprimDateDossier'=>$hprim['dateDossier'],
-      'hprimAllSerialize'=>serialize($hprim),
-      'pjNombre'=>count($pj),
-      'pjSerializeName'=>serialize($pj)
+    $data=array(
+    'txtFileName'=>$file,
+    'txtDatetime'=>$filedata['datetime'],
+    'txtNumOrdre'=>$filedata['numOrdre'],
+    'hprimIdentite'=>$hprim['prenom'].' '.$hprim['nom'],
+    'hprimExpediteur'=>$hprim['expediteur'],
+    'hprimCodePatient'=>$hprim['codePatient'],
+    'hprimDateDossier'=>$hprim['dateDossier'],
+    'hprimAllSerialize'=>serialize($hprim),
+    'pjNombre'=>count($pj),
+    'pjSerializeName'=>serialize($pj)
     );
 
-      msSQL::sqlInsert('inbox', $data);
-      echo mysqli_error($mysqli);
+    msSQL::sqlInsert('inbox', $data);
+    echo mysqli_error($mysqli);
   }
 }
